pyproject.toml
[tool.poetry]
name = "hbs-report-generator"
version = "0.1.0"
description = "hbs-report-generator"
authors = ["Roman Gabert <graa@list.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
pydantic = "^2.8.2"
tenacity = "^9.0.0"
requests-toolbelt = "^1.0.0"
jinja2 = "^3.1.4"
responses = "^0.25.3"
pydantic-settings = "^2.4.0"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
pyproject-flake8 = "^7.0.0"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
skip-string-normalization = "true"

[tool.flake8]
max-line-length = 100
